+++ src/main/java/com/grepp/coffee/service/menu/MenuService.java
@@ -1,10 +1,11 @@
 package com.grepp.coffee.service.menu;
 
-import com.grepp.coffee.app.model.coffee.dto.CoffeeDto;
+import com.grepp.coffee.app.model.dto.CoffeeDto;
+import java.util.List;
 
 public interface MenuService {
     void addMenu(CoffeeDto coffeeDto);  // 커피 메뉴 추가
-    CoffeeDto getCoffee();
+    List<CoffeeDto> getAllCoffee();
     void updateMenu(CoffeeDto coffeeDto);
     void deleteMenu(CoffeeDto coffeeDto);
 }
