+++ src/main/java/com/grepp/coffee/service/order/OrderService.java
@@ -1,6 +1,6 @@
 package com.grepp.coffee.service.order;
 
-import com.grepp.coffee.app.model.detailedorder.dto.DetailedOrderDto;
+import com.grepp.coffee.app.model.dto.DetailedOrderDto;
 
 public interface OrderService {
     void processOrder(DetailedOrderDto detailedOrderDto);
